assistant_coach/requirements.txt
groq>=0.9.0
streamlit>=1.35.0
numpy>=2.0.0
pandas>=2.2.0
requests>=2.31.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
